requirements.txt
asgiref
certifi
cloudinary
dj-database-url
Django
django-cors-headers
django-filter
djangorestframework
djangorestframework-simplejwt
Markdown
pip
psycopg2-binary
PyJWT
python-dotenv
pytz
six
sqlparse
typing_extensions
tzdata
urllib3
